feat: 初步支持1.21.11 fabric

+++ src/main/java/com/github/ysbbbbbb/kaleidoscopecookery/mixin/FarmBlockMixin.java
@@ -20,8 +20,8 @@ public class FarmBlockMixin {
             target = "Lnet/minecraft/world/level/block/FarmBlock;turnToDirt(Lnet/minecraft/world/entity/Entity;Lnet/minecraft/world/level/block/state/BlockState;Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;)V",
                     shift = At.Shift.BEFORE),
             cancellable = true)
-    private void onFallOn(Level level, BlockState state, BlockPos pos, Entity entity, float fallDistance, CallbackInfo ci) {
-        FarmlandTrampleEvent farmlandTrampleEvent = new FarmlandTrampleEvent(level, pos, state, fallDistance, entity);
+    private void onFallOn(Level level, BlockState state, BlockPos pos, Entity entity, double fallDistance, CallbackInfo ci) {
+        FarmlandTrampleEvent farmlandTrampleEvent = new FarmlandTrampleEvent(level, pos, state, (float) fallDistance, entity);
         ModEvents.FARMLAND_TRAMPLE.invoker().onFarmlandTrample(farmlandTrampleEvent);
         if (farmlandTrampleEvent.isCanceled()) ci.cancel();
